refactor(nebula_ros): move mt_queue to common
